Add torch dtype.

+++ src/LLM/LLM_train.py
@@ -20,7 +20,7 @@ MAX_LENGTH = 1024
 
 def model_small_lm_360m() -> tuple[AutoModelForCausalLM, AutoTokenizer]:
     checkpoint = "HuggingFaceTB/SmolLM-360M-Instruct"
-    model = AutoModelForCausalLM.from_pretrained(checkpoint).to(device)
+    model = AutoModelForCausalLM.from_pretrained(checkpoint, torch_dtype=torch.float16).to(device)
     tokenizer = AutoTokenizer.from_pretrained(checkpoint)
     add_special_tokens_if_needed(tokenizer, model)
     return model, tokenizer
@@ -28,7 +28,7 @@ def model_small_lm_360m() -> tuple[AutoModelForCausalLM, AutoTokenizer]:
 
 def model_small_lm_135m() -> tuple[AutoModelForCausalLM, AutoTokenizer]:
     checkpoint = "HuggingFaceTB/SmolLM-135M"
-    model = AutoModelForCausalLM.from_pretrained(checkpoint).to(device)
+    model = AutoModelForCausalLM.from_pretrained(checkpoint, torch_dtype=torch.float16).to(device)
     tokenizer = AutoTokenizer.from_pretrained(checkpoint)
     add_special_tokens_if_needed(tokenizer, model)
     return model, tokenizer
@@ -36,7 +36,7 @@ def model_small_lm_135m() -> tuple[AutoModelForCausalLM, AutoTokenizer]:
 
 def model_small_lm_1b() -> tuple[AutoModelForCausalLM, AutoTokenizer]:
     checkpoint = "HuggingFaceTB/SmolLM-1.7B"
-    model = AutoModelForCausalLM.from_pretrained(checkpoint).to(device)
+    model = AutoModelForCausalLM.from_pretrained(checkpoint, torch_dtype=torch.float16).to(device)
     tokenizer = AutoTokenizer.from_pretrained(checkpoint)
     add_special_tokens_if_needed(tokenizer, model)
     return model, tokenizer
